Ajustes no css da tela de cadastro do usuário e outras pqns alterações
