+++ src/App.css
@@ -1,8 +1,8 @@
 .App {
   text-align: center;
-  display: flex;
+  /* display: flex;
   flex-direction: column;
-  justify-content: space-between;
+  justify-content: space-between; */
   min-height: 100vh;
 }
