+++ src/styles/global.js
@@ -1,23 +1,24 @@
-import { createGlobalStyle } from "styled-components";
-import px2vw from "../utils/px2vw";
+import  styled from "styled-components";
+import px2vw from "../hooks/px2vw";
 
-export const Global = createGlobalStyle`
+export const Global = styled.div`
   * {
     margin: 0;
     padding: 0;
     box-sizing: border-box;
   }
-  :root {
-      font-size: ${px2vw(24)};
+  // :root {
+  //     font-size: ${px2vw(24)};
 
-      @media (min-width: 768px) {
-        font-size: ${px2vw(16)};
-      }
+  //     @media (min-width: 768px) {
+  //       font-size: ${px2vw(16)};
+  //     }
 
-      @media (min-width: 1024px) {
-        font-size: ${px2vw(14)};
-      }
-    }
-`;
-
-export default Global;
+  //     @media (min-width: 1024px) {
+  //       font-size: ${px2vw(14)};
+  //     }
+  //   }
+`
+export const ContentContainer = styled.div`
+    min-height: 75vh;
+`
